+++ src/test/java/com/intive/exercise/rent/RentServiceImplTest.java
@@ -1,0 +1,28 @@
+package com.intive.exercise.rent;
+
+import com.intive.exercise.rent.models.Rent;
+import com.intive.exercise.rent.models.RentType;
+import com.intive.exercise.rent.parameters.StrategyRentType;
+import com.intive.exercise.rent.services.RentService;
+import com.intive.exercise.rent.services.strategies.*;
+import com.intive.exercise.user.User;
+import com.intive.exercise.utils.Constants;
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.junit.Assert.*;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class RentServiceImplTest {
+
+
+
+
+}
